pyproject.toml
[project]
name = "user-notification-system"
version = "0.1.0"
description = ""
authors = [
    {name = "breynisson",email = "bjorgvin.reynisson@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio (>=1.69.0,<2.0.0)",
    "grpcio-tools (>=1.69.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "user_notification_system" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
hypothesis = "^6.124.2"

[tool.pytest.ini_options]
pythonpath = [
    ".", "generated"
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.poetry.scripts]
generate-grpc = "user_notification_system.generate_grpc:generate_protos"
start-server = "user_notification_system.server:main"

